decouple Track encoding from UploadManager -> new TrackEncoder class
